script: switch to keeping data in associative arrays
There's a bunch of data being processed, so it's not a bad idea to
structure it a little better than a bunch of environment variables.

This commit stores the data in an associative array with keys based
on the data being requested and the the day it's being requested for.

It makes the code a little harder to read, but doing it this way let's
us keep the data around after it's processed so we can look back at
it (which might later help with the FIXME we have in the code)