script: fix summary stat line on bottom if script run early in morning
After printing the last N days of reports, a summary line with averages
is printed.

This summary line relies on today's report being available.  It uses
today's report to get the totals it needs when computing the averages.

But if the script is run in the morning, today's report won't be
available yet.

This commit changes the summary line to instead look at the last
report available when figuring out the average.