script: factor out repetitive code into a function
Since bash doesn't support multidimensional arrays, and we have
two axes (day number and stat name), we encode key names based
on both axes into a single global data store.

In order to abstract out the encoding scheme used for key names
from all the users of the array, we maintain an associative array
of key names for the current day (and for the first day and for
the last day), that takes the stat name as a key, and
returns the stat name encoded with the day.

The problem is, generating these arrays of key names requires
looping over all the stat names we are going to be accessing.
Since we generate 3 arrays, we need that loop 3 times in the
code.

This moves the loops to a function to reduce code duplication.